All javascript is now written as AMD modules
