+++ views/default/file_tools/list/files.php
@@ -10,11 +10,11 @@ $offset = (int) elgg_extract("offset", $vars, 0);
 $folder_content = "";
 if (empty($offset)) {
 	$folder_content = elgg_view("file_tools/breadcrumb", array("entity" => $folder));
-	
+
 	if (!($sub_folders = file_tools_get_sub_folders($folder))) {
 		$sub_folders = array();
 	}
-	
+
 	$entities = array_merge($sub_folders, $files);
 } else {
 	$entities = $files;
@@ -25,11 +25,11 @@ if (!empty($entities)) {
 		"full_view" => false,
 		"pagination" => false
 	);
-	
+
 	elgg_push_context("file_tools_selector");
-	
+
 	$files_content = elgg_view_entity_list($entities, $params);
-	
+
 	elgg_pop_context();
 }
 
@@ -51,22 +51,22 @@ if (empty($files_content)) {
 		}
 		$files_content .= "</div>";
 	}
-	
+
 	// only show selectors on the first load
 	if (empty($offset)) {
 		$files_content .= "<div class='clearfix'>";
-		
+
 		if (elgg_get_page_owner_entity()->canEdit()) {
 			$files_content .= '<a id="file_tools_action_bulk_delete" href="javascript:void(0);">' . elgg_echo("file_tools:list:delete_selected") . '</a> | ';
 		}
-		
+
 		$files_content .= "<a id='file_tools_action_bulk_download' href='javascript:void(0);'>" . elgg_echo("file_tools:list:download_selected") . "</a>";
-		
+
 		$files_content .= "<a id='file_tools_select_all' class='float-alt' href='javascript:void(0);'>";
 		$files_content .= "<span>" . elgg_echo("file_tools:list:select_all") . "</span>";
 		$files_content .= "<span class='hidden'>" . elgg_echo("file_tools:list:deselect_all") . "</span>";
 		$files_content .= "</a>";
-		
+
 		$files_content .= "</div>";
 	}
 }
@@ -81,16 +81,6 @@ echo "</div>";
 
 $page_owner = elgg_get_page_owner_entity();
 
-if ($page_owner->canEdit() || (elgg_instanceof($page_owner, "group") && $page_owner->isMember())) { ?>
-<script type="text/javascript">
-
-	$(function(){
-		
-		elgg.file_tools.initialize_file_draggable();
-		elgg.file_tools.initialize_folder_droppable();
-		
-	});
-
-</script>
-<?php
+if ($page_owner->canEdit() || (elgg_instanceof($page_owner, "group") && $page_owner->isMember())) {
+	elgg_require_js('file_tools/site');
 }
